+++ mayan/__init__.py
@@ -1,7 +1,7 @@
 __title__ = 'Mayan EDMS'
 __version__ = '4.3b1'
 __build__ = 0x040300
-__build_string__ = 'v4.2.7-228-g0865963913_Thu Jul 21 23:59:57 2022 -0400'
+__build_string__ = 'v4.2.8-231-g6d422ce4e6_Sun Jul 24 18:02:12 2022 -0400'
 __django_version__ = '3.2'
 __author__ = 'Roberto Rosario'
 __author_email__ = 'roberto.rosario@mayan-edms.com'